Splitted Flask_server into module related to website sections
